requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
firebase-admin==6.5.0
pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
